Deal with directories and symlinks without treating them as missing
Using backup.getFileName for directory or symlink entries would throw
an exception due to the file missing, so we never got to stat the
file and check if it was a directory.

By using the stat mode from the backup itself we can skip both
directories and symlinks before we get to the backup.getFileName
step.

Ideally we'd restore symlinks too, but this is left for future
work. Restoring the mode of the files probably doesn't add much,
but in theory we could do that too.